AI/AI Reports/Analytics/requirements.txt
pymongo==4.6.1
certifi==2023.11.17
openai==1.7.2
python-dotenv==1.0.0
sentence-transformers==2.2.2
